Update to work with NodeJS 20, rip out some junk
I have successfully gotten it to work with NodeJS 20.  I had to rip out
Firebase, a redirect plugin, and some actions associated with Firebase
in order to get the "Start Captioning" button to work at all.  But it
seems to be working now!

Up next: rip out as much of this as possible so that the process of
updating dependencies doesn't suck so much.